requirements.txt
fastmcp
httpx
lyricsgenius
python-dotenv
pytest
pytest-asyncio
